Update tools.py
changed file operations to avoid processor hang. extron qxi firmware 1.11.0000-b0006 broke extronlib.system.File

+++ release/tools.py
@@ -220,6 +220,9 @@ Changelog:
         - fixed intellisense recognition of non-list parameters for VirtualUI class
     v 1.8.0.10 - tools.py modification
         - fixed issue on knob motion event producing an error in the log. the callback still executed anyway, though.
+    v 1.8.0.11 - tools.py modification
+        - changed file operations to avoid processor hang. extron qxi firmware 1.11.0000-b0006 broke extronlib.system.File
+            - when attempting to open a nonexistent file, the processor hangs indefinitely.
 """
 
 
@@ -386,20 +389,22 @@ class NonvolatileValues():
     def ReadValues(self):
         f = None #type:_File
         values = {}
-        with _File(self.__filename,'r') as f:
-            if f:
-                try:
-                    values = _json.load(f)
-                except Exception as e:
-                    if sys_allowed_flag:
-                        err_msg = 'EXCEPTION:{}:{}:{}:deleting corrupted file'.format(__class__.__name__,sys._getframe().f_code.co_name,traceback.format_exc())
-                    else:
-                        err_msg = 'EXCEPTION:{}:{}:{}:deleting corrupted file'.format(__class__.__name__,'ReadValues',e)
-                    print(err_msg)
-                    DebugPrint.Print(err_msg)
-                    _ProgramLog(err_msg)
-                    _File.DeleteFile(self.__filename)
-                f.close()
+        if _File.Exists(self.__filename):
+            with _File(self.__filename,'r') as f:
+                if f:
+                    try:
+                        values = _json.load(f)
+                    except Exception as e:
+                        if sys_allowed_flag:
+                            err_msg = 'EXCEPTION:{}:{}:{}:deleting corrupted file'.format(__class__.__name__,sys._getframe().f_code.co_name,traceback.format_exc())
+                        else:
+                            err_msg = 'EXCEPTION:{}:{}:{}:deleting corrupted file'.format(__class__.__name__,'ReadValues',e)
+                        print(err_msg)
+                        DebugPrint.Print(err_msg)
+                        _ProgramLog(err_msg)
+                        _File.DeleteFile(self.__filename)
+                    f.close()
+        DebugPrint.Print('rv 2')
         self.values = values
         if self.__syncvaluesfunctions:
             for f in self.__syncvaluesfunctions:
@@ -418,19 +423,20 @@ class NonvolatileValues():
 
     def SaveValues(self):
         f = None #type:_File
-        with _File(self.__filename,'w') as f:
-            if f:
-                try:
-                    _json.dump(self.values,f)
-                except Exception as e:
-                    if sys_allowed_flag:
-                        err_msg = 'EXCEPTION:{}:{}:{}'.format(__class__.__name__,sys._getframe().f_code.co_name,traceback.format_exc())
-                    else:
-                        err_msg = 'EXCEPTION:{}:{}:{}'.format(__class__.__name__,'SaveValues',e)
-                    print(err_msg)
-                    DebugPrint.Print(err_msg)
-                    _ProgramLog(err_msg)
-                f.close()
+        if _File.Exists(self.__filename):
+            with _File(self.__filename,'w') as f:
+                if f:
+                    try:
+                        _json.dump(self.values,f)
+                    except Exception as e:
+                        if sys_allowed_flag:
+                            err_msg = 'EXCEPTION:{}:{}:{}'.format(__class__.__name__,sys._getframe().f_code.co_name,traceback.format_exc())
+                        else:
+                            err_msg = 'EXCEPTION:{}:{}:{}'.format(__class__.__name__,'SaveValues',e)
+                        print(err_msg)
+                        DebugPrint.Print(err_msg)
+                        _ProgramLog(err_msg)
+                    f.close()
 
 
     def AddSyncValuesFunction(self,func):
@@ -455,8 +461,11 @@ class PasswordManager():
         self.__password_numeric = password_numeric
         self.__character_count = character_count
         self.__backdoor_password = backdoor_password
+        DebugPrint.Print("pwm 1")
         self.__nvram = NonvolatileValues('{}.csv'.format(manager_id))
+        DebugPrint.Print("pwm 2")
         self.__nvram.ReadValues()
+        DebugPrint.Print("pwm 3")
 
     def GetPassword(self,password_id=''):
         try:
@@ -540,31 +549,34 @@ class ProgramLogSaver():
     def __readdummyprogramlog():
         f = None #type:_File
         log = None
-        with _File('/ProgramLogs/temp.log','r') as f:
-            if f:
-                try:
-                    log = f.read()
-                except Exception as e:
-                    if sys_allowed_flag:
-                        err_msg = 'EXCEPTION:{}:{}:{}'.format(__class__.__name__,sys._getframe().f_code.co_name,traceback.format_exc())
-                    else:
-                        err_msg = 'EXCEPTION:{}:{}:{}'.format(__class__.__name__,'__readdummyprogramlog',e)
-                    print(err_msg)
-                    DebugPrint.Print(err_msg)
-                    _ProgramLog(err_msg)
-                f.close()
+        if _File.Exists('/ProgramLogs/temp.log'):
+            with _File('/ProgramLogs/temp.log','r') as f:
+                if f:
+                    try:
+                        log = f.read()
+                    except Exception as e:
+                        if sys_allowed_flag:
+                            err_msg = 'EXCEPTION:{}:{}:{}'.format(__class__.__name__,sys._getframe().f_code.co_name,traceback.format_exc())
+                        else:
+                            err_msg = 'EXCEPTION:{}:{}:{}'.format(__class__.__name__,'__readdummyprogramlog',e)
+                        print(err_msg)
+                        DebugPrint.Print(err_msg)
+                        _ProgramLog(err_msg)
+                    f.close()
         return log
 
 
     def __saveprogramlog():
-        with _File(ProgramLogSaver.__filename, 'w') as f:
-            if f:
-                _SaveProgramLog(f)
+        if _File.Exists(ProgramLogSaver.__filename):
+            with _File(ProgramLogSaver.__filename, 'w') as f:
+                if f:
+                    _SaveProgramLog(f)
 
     def __savedummyprogramlog():
-        with _File('/ProgramLogs/temp.log', 'w') as f:
-            if f:
-                _SaveProgramLog(f)
+        if _File.Exists('/ProgramLogs/temp.log'):
+            with _File('/ProgramLogs/temp.log', 'w') as f:
+                if f:
+                    _SaveProgramLog(f)
 
 
     def __checkprogramlog(timer,count):
@@ -678,19 +690,20 @@ class DebugFileLogSaver():
                 if len(DebugFileLogSaver.__cur_logs) > 1:
                     filename = DebugFileLogSaver.__getfilename()
                     f = None #type:_File
-                    with _File(filename, 'a') as f:
-                        if f:
-                            try:
-                                f.write('\n'.join(DebugFileLogSaver.__cur_logs))
-                            except Exception as e:
-                                if sys_allowed_flag:
-                                    err_msg = 'EXCEPTION:{}:{}:{}'.format(__class__.__name__,sys._getframe().f_code.co_name,traceback.format_exc())
-                                else:
-                                    err_msg = 'EXCEPTION:{}:{}:{}'.format(__class__.__name__,'__checklogloop',e)
-                                print(err_msg)
-                                DebugPrint.Print(err_msg)
-                                _ProgramLog(err_msg)
-                            f.close()
+                    if _File.Exists(filename):
+                        with _File(filename, 'a') as f:
+                            if f:
+                                try:
+                                    f.write('\n'.join(DebugFileLogSaver.__cur_logs))
+                                except Exception as e:
+                                    if sys_allowed_flag:
+                                        err_msg = 'EXCEPTION:{}:{}:{}'.format(__class__.__name__,sys._getframe().f_code.co_name,traceback.format_exc())
+                                    else:
+                                        err_msg = 'EXCEPTION:{}:{}:{}'.format(__class__.__name__,'__checklogloop',e)
+                                    print(err_msg)
+                                    DebugPrint.Print(err_msg)
+                                    _ProgramLog(err_msg)
+                                f.close()
                     print('DebugFileLogSaver:added {} logs to file:{}'.format(len(DebugFileLogSaver.__cur_logs),filename))
                     DebugFileLogSaver.__cur_logs = [DebugFileLogSaver.__init_log]
     __save_wait = _Wait(1,__checklogloop)
@@ -5201,9 +5214,10 @@ class ProcessorDeviceWrapper(DebugServer):
         self.__printToLog(self.__friendly_name,'Command','SaveProgramLog')
         dt = _datetime.now()
         filename = 'ProgramLog {}.txt'.format(dt.strftime('%Y-%m-%d %H%M%S'))
-        with _File(filename, 'w') as f:
-            if f:
-                _SaveProgramLog(f)
+        if _File.Exists(filename):
+            with _File(filename, 'w') as f:
+                if f:
+                    _SaveProgramLog(f)
 
 
     def __printToLog(self,device,message_type,data):
